# Avoid readByte and appendByte in nulls to bits serde