```
AFTER

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLineitem      avgt   60  1,173 ± 0,019  ns/op
BenchmarkBlockSerde.deserializeLongWithNull  avgt   60  2,286 ± 0,032  ns/op
BenchmarkBlockSerde.serializeLineitem        avgt   60  3,512 ± 0,039  ns/op
BenchmarkBlockSerde.serializeLongWithNull    avgt   60  2,425 ± 0,030  ns/op

BEFORE

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLineitem      avgt   60  1,187 ± 0,015  ns/op
BenchmarkBlockSerde.deserializeLongWithNull  avgt   60  2,474 ± 0,022  ns/op
BenchmarkBlockSerde.serializeLineitem        avgt   60  3,579 ± 0,046  ns/op
BenchmarkBlockSerde.serializeLongWithNull    avgt   60  3,252 ± 0,030  ns/op
```
1/3 improvement for `serializeLongWithNull`, 10% improvement for `deserializeLongWithNull`