@sopel39 2020-11-13
> So the conclusion is that handling data through Slices suck.

* Bound checks are an issue
* Unsafe itself is not as fast as native arrays. 
* Bulk byte operations is faster than byte-by-byte operations.